requirements.txt
# PDF Processing (CORE)
PyPDF2==3.0.1
reportlab==4.0.7
Pillow==10.1.0
PyMuPDF==1.23.8

# Azure SDK (REQUIRED)
azure-functions==1.18.0
azure-storage-blob==12.19.0
azure-identity==1.15.0

# HTTP & Validation (CORE)
aiohttp==3.9.1
pydantic==2.5.0

# Testing (DEV)
pytest==7.4.3
pytest-asyncio==0.21.1